+++ src/renderer/components/wallet/assets/AssetDetails.tsx
@@ -238,7 +238,7 @@ export const AssetDetails: React.FC<Props> = (props): JSX.Element => {
           </TextButton>
         </Col>
         <Col span={24}>
-          {asset.synth ? (
+          {asset.synth || AssetRuneNative ? (
             <WarningView
               subTitle={intl.formatMessage(
                 { id: 'wallet.txs.history.disabled' },