+++ src/renderer/i18n/fr/settings.ts
@@ -22,6 +22,8 @@ const settings: SettingMessages = {
   'setting.wallet.hdpath.legacy.info': 'Chemin de dérivation Legacy {path}',
   'setting.wallet.hdpath.ledgerlive.info': 'Chemin de dérivation Ledger Live {path}',
   'setting.wallet.hdpath.metamask.info': 'Chemin de dérivation Metamask {path}',
+  'setting.wallet.storeAddress.info':
+    "Enregistrez vos adresses de confiance dans le stockage local, elles seront répertoriées sous la chaîne ou dans la vue d'envoi",
   'setting.thornode.node.error.unhealthy': "L'API THORNode semble défaillante en vérifiant {endpoint}",
   'setting.thornode.node.error.url': "URL de l'API THORNode invalide. Veuillez vérifier et réessayer",
   'setting.thornode.rpc.error.url': 'URL RPC THORNode invalide. Veuillez vérifier et réessayer',