+++ src/renderer/services/storage/index.ts
@@ -1,7 +1,8 @@
 import * as common from './common'
 import * as pools from './pools'
+import * as userAddresses from './userAddresses'
 import * as userChains from './userChains'
 import * as userAssets from './userChainTokens'
 import * as userNodes from './userNodes'
 
-export { common, userNodes, pools, userChains, userAssets }
+export { common, userNodes, pools, userChains, userAddresses, userAssets }