+++ src/setupTests.ts
@@ -57,6 +57,7 @@ global.window.apiHDWallet = { ...mockApi.apiHDWallet }
 global.window.apiCommonStorage = { ...mockApi.apiCommonStorage }
 global.window.apiUserNodesStorage = { ...mockApi.apiUserNodesStorage }
 global.window.apiChainStorage = { ...mockApi.apiChainStorage }
+global.window.apiAddressStorage = { ...mockApi.apiAddressStorage }
 global.window.apiAssetStorage = { ...mockApi.apiAssetStorage }
 global.window.apiPoolsStorage = { ...mockApi.apiPoolsStorage }