+++ src/shared/const.ts
@@ -1,6 +1,12 @@
 import { DEFAULT_USER_ASSETS } from '../renderer/const'
 import { PoolsStorageEncoded } from './api/io'
-import { StoreFilesContent, UserAssetStorage, UserChainStorage, UserNodesStorage } from './api/types'
+import {
+  StoreFilesContent,
+  UserAssetStorage,
+  UserChainStorage,
+  UserNodesStorage,
+  UserTrustedAddressStorage
+} from './api/types'
 import { DEFAULT_EVM_HD_MODE } from './evm/types'
 import { DEFAULT_LOCALE } from './i18n/const'
 import { DEFAULT_MAYANODE_API_URLS, DEFAULT_MAYANODE_RPC_URLS } from './mayachain/const'
@@ -53,6 +59,13 @@ export const CHAINS_STORAGE_DEFAULT: UserChainStorage = {
   version: CHAINS_STORAGE_VERSION,
   chains: Object.keys(DEFAULT_ENABLED_CHAINS) as EnabledChain[]
 }
+
+const ADDRESS_STORAGE_VERSION = '1'
+
+export const ADDRESS_STORAGE_DEFAULT: UserTrustedAddressStorage = {
+  version: ADDRESS_STORAGE_VERSION,
+  addresses: []
+}
 /// increase it by `1` if you want to ignore previous version of `common` storage
 const ASSETS_STORAGE_VERSION = '2'
 
@@ -91,6 +104,7 @@ export const DEFAULT_STORAGES: StoreFilesContent = {
     mayanodeRpc: DEFAULT_MAYANODE_RPC_URLS
   },
   userChains: CHAINS_STORAGE_DEFAULT,
+  userAddresses: ADDRESS_STORAGE_DEFAULT,
   userAssets: ASSETS_STORAGE_DEFAULT,
   userNodes: USER_NODES_STORAGE_DEFAULT,
   pools: POOLS_STORAGE_DEFAULT